@@ -1,7 +1,7 @@
-import { Module } from '@nestjs/common';
-import { RestApiService } from './rest-api.service';
-import { RestApiController } from './rest-api.controller';
-import { JwtModule } from '@nestjs/jwt';
+import { Module } from '@nestjs/common'
+import { RestApiService } from './rest-api.service'
+import { RestApiController } from './rest-api.controller'
+import { JwtModule } from '@nestjs/jwt'
 
 @Module({
   imports: [JwtModule.register({
