@@ -7,7 +7,7 @@ import { SERVER_LINK } from "../../helpers/config";
 import axios from "axios";
 import { useRouter } from "next/router";
 
-const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
+const OrganizationModal = ({ modal, setModal,editForm , organizationId , orgData }) => {
   //For Image Preview
   const [selectedImage, setSelectedImage] = useState();
   const router = useRouter();
@@ -26,40 +26,40 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
 
   const { register, handleSubmit } = useForm();
 
-  useEffect(()=>{
+  // useEffect(()=>{
 
-    if(orgData){
+  //   if(orgData){
 
-      userData();
-    }
-  },[editForm])
+  //     userData();
+  //   }
+  // },[editForm])
 
 
- function userData(){
+//  function userData(){
 
-    // setButtonText("Update");/
-    // setEditForm(true);
-    // setOrganizationId(org_id);
+//     // setButtonText("Update");/
+//     // setEditForm(true);
+//     // setOrganizationId(org_id);
 
-    // first find the user with the id
+//     // first find the user with the id
    
-    console.log("this is the orgdata "+orgData);
-    console.log(orgData);
+//     console.log("this is the orgdata "+orgData);
+//     console.log(orgData);
 
-        setName(orgData.name);
-        setEmail(orgData.email);
-        setMobile(orgData.mobile);
-        setState(orgData.state);
-        setAddress(orgData.address);
-        setCity(orgData.city);
-        setPincode(orgData.pincode);
-        setQuota(orgData.quota);
-        setPassword(orgData.password)
-        setButtonText("Update");
+//         setName(orgData.name);
+//         setEmail(orgData.email);
+//         setMobile(orgData.mobile);
+//         setState(orgData.state);
+//         setAddress(orgData.address);
+//         setCity(orgData.city);
+//         setPincode(orgData.pincode);
+//         setQuota(orgData.quota);
+//         setPassword(orgData.password)
+//         setButtonText("Update");
 
      
 
-}
+// }
   // This function will be triggered when the file field change
   const imageChange = (e) => {
     if (e.target.files && e.target.files.length > 0) {
@@ -72,18 +72,28 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
     setSelectedImage();
   };
 
-  useEffect(() => {
-    if (!modal) {
-      setSelectedImage();
-    }
-  }, [modal]);
+  // useEffect(() => {
+  //   if (!modal) {
+  //     setSelectedImage();
+  //   }
+  // }, [modal]);
 
   // for sending the data to the backend
   const checkWithDatabase = async (data) => {
     console.log("This is thge data ");
-    console.log(data);
     data.status = true;
-    data = JSON.stringify(data);
+    data.name = name
+    data.email = email
+    data.mobile = mobile
+    data.password = password
+    data.city = city
+    data.state = state
+    data.pincode = pincode
+    data.address = address
+    data.quota = quota
+    let OrganizationData = JSON.stringify(data);
+    console.log(data);
+
 
     // for taking the patch api data
     if (editForm) {
@@ -112,7 +122,7 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
           Accept: "application/json",
           "Content-Type": "application/json;charset=UTF-8",
         },
-        data,
+        data : OrganizationData,
       })
         .then((response) => {
           router.replace(router.asPath);
@@ -127,10 +137,20 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
   return (
     <>
       <PureModal
-        //header={<div className="bg-blue-600 p-2 font-bold text-lg text-center text-white">Category</div>}
+        //header={<div classNameName="bg-blue-600 p-2 font-bold text-lg text-center text-white">Category</div>}
         isOpen={modal}
         width="800px"
         onClose={() => {
+          setName("")
+          setEmail("")
+          // setPassword("") 
+          setPassword("") 
+          setCity("")
+          setState("")
+          setPincode("")
+          setAddress("")
+          setMobile("")
+          setQuota("")
           setModal(false);
           return true;
         }}
@@ -140,198 +160,191 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
             <p>{buttonText} Organization</p>
           </div>
 
-          <div class="py-6 px-6 lg:px-8">
+          <div className="py-6 px-6 lg:px-8">
             <form
-              class="w-full max-w-lg"
+              className="w-full max-w-lg"
               onSubmit={handleSubmit((data) => checkWithDatabase(data))}
             >
-              <div class="flex flex-wrap -mx-3 mb-6">
-                <div class="w-full md:w-1/2 px-3 mb-6 md:mb-0">
+              <div className="flex flex-wrap -mx-3 mb-6">
+                <div className="w-full md:w-1/2 px-3 mb-6 md:mb-0">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-first-name"
                   >
                     Name
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border  rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border  rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
                     id="grid-first-name"
                     type="text"
                     value={name}
-                    {...register("name", {
-                      onChange: (e) => setName(e.target.value)
-                    })}
+                    required = "required"
+                    onChange = { (e) => setName(e.target.value)}
                     placeholder="Jane"
                   />
-                  {/* <p class="text-red-500 text-xs italic">
+                  {/* <p className="text-red-500 text-xs italic">
                     Please fill out this field.   property - > border-red-500
                   </p> */}
                 </div>
-                <div class="w-full md:w-1/2 px-3">
+                <div className="w-full md:w-1/2 px-3">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-last-name"
                   >
                     Email
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-email"
                     type="email"
                     placeholder="example@gmail.com "
                     value={email}
-                    {...register("email", {
-                      onChange: (e) => setEmail(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setEmail(e.target.value)}
                   />
                 </div>
               </div>
 
-              <div class="flex flex-wrap -mx-3 mb-6">
-                <div class="w-full px-3">
+              <div className="flex flex-wrap -mx-3 mb-6">
+                <div className="w-full px-3">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-password"
                   >
                     Password
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-password"
                     type="password"
                     placeholder="******************"
                     value={password}
-                    {...register("password", {
-                      onChange: (e) => setPassword(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setPassword(e.target.value)}
                   />
-                  <p class="text-gray-600 text-xs italic">
+                  <p className="text-gray-600 text-xs italic">
                     Make it as long and as crazy as you'd like
                   </p>
                 </div>
               </div>
 
-              <div class="flex flex-wrap -mx-3 mb-2">
-                <div class="w-full md:w-1/3 px-3 mb-6 md:mb-0">
+              <div className="flex flex-wrap -mx-3 mb-2">
+                <div className="w-full md:w-1/3 px-3 mb-6 md:mb-0">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-city"
                   >
                     City
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-city"
                     type="text"
                     placeholder="Albuquerque"
                     value={city}
-                    {...register("city", {
-                      onChange: (e) => setCity(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setCity(e.target.value)}
                   />
                 </div>
 
-                <div class="w-full md:w-1/3 px-3 mb-6 md:mb-0">
+                <div className="w-full md:w-1/3 px-3 mb-6 md:mb-0">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-city"
                   >
                     State
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-state"
                     type="text"
                     placeholder="State"
                     value={state}
-                    {...register("state", {
-                      onChange: (e) => setState(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setState(e.target.value)}
                   />
                 </div>
-                <div class="w-full md:w-1/3 px-3 mb-6 md:mb-0">
+                <div className="w-full md:w-1/3 px-3 mb-6 md:mb-0">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-zip"
                   >
                     Pin Code
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-zip"
                     type="text"
                     placeholder="90210"
                     value={pincode}
-                    {...register("pincode", {
-                      onChange: (e) => setPincode(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setPincode(e.target.value)}
                   />
                 </div>
               </div>
-              <div class="flex flex-wrap -mx-3 mb-6">
-                <div class="w-full px-3">
+              <div className="flex flex-wrap -mx-3 mb-6">
+                <div className="w-full px-3">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-password"
                   >
                     Address
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-address"
                     type="text"
                     placeholder="Your Office number... "
                     value={address}
-                    {...register("address", {
-                      onChange: (e) => setAddress(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setAddress(e.target.value)}
                   />
                 </div>
               </div>
-              <div class="flex flex-wrap -mx-3 mb-6">
-                <div class="w-full md:w-1/2 px-3 mb-6 md:mb-0">
+              <div className="flex flex-wrap -mx-3 mb-6">
+                <div className="w-full md:w-1/2 px-3 mb-6 md:mb-0">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-first-name"
                   >
                     Mobile
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border  rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border  rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
                     id="grid-mobile"
                     type="text"
                     placeholder="+91 "
                     value={mobile}
-                    {...register("mobile", {
-                      onChange: (e) => setMobile(e.target.value),
-                    })}
+                    required = "required"
+                    onChange = { (e) => setMobile(e.target.value)}
                   />
-                  {/* <p class="text-red-500 text-xs italic">
+                  {/* <p className="text-red-500 text-xs italic">
                     Please fill out this field.   property - > border-red-500
                   </p> */}
                 </div>
-                <div class="w-full md:w-1/2 px-3">
+                <div className="w-full md:w-1/2 px-3">
                   <label
-                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
                     for="grid-last-name"
                   >
                     Quota
                   </label>
                   <input
-                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-quota"
                     type="text"
                     placeholder="e.g. 1000"
                     value={quota}
-                    {...register("quota", {
-                      onChange: (e) => setQuota(e.target.value),
-                    })}
+                    required = "required"
+                    // {...register("quota", {
+                      onChange = { (e) => setQuota(e.target.value)}
+                    // })}
                   />
                 </div>
               </div>
               <button
                 type="submit"
-                class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+                className="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
               >
                 {buttonText}
               </button>
@@ -341,9 +354,9 @@ const Modal = ({ modal, setModal,editForm , organizationId , orgData }) => {
           {/* */}
         </div>
       </PureModal>
-      ;
+      
     </>
   );
 };
 
-export default Modal;
+export default OrganizationModal;
