@@ -1,8 +1,15 @@
 import React, { useState } from "react";
-import Modal from "./Modal";
-import ParticipantModal from './ParticipantModal';
+import OrganizationModal from "./Modal";
+import ParticipantModal from "./ParticipantModal";
+import LevelModal from "./LevelModal";
+import ModuleModal from "./ModuleModal";
 
-const PageComponentTitle = ({ title, titleDescription, buttonTitle, editForm }) => {
+const PageComponentTitle = ({
+  title,
+  titleDescription,
+  buttonTitle,
+  editForm,
+}) => {
   const [modal, setModal] = useState(false);
   // const [editForm,setEditForm] = useState(false);
 
@@ -11,6 +18,7 @@ const PageComponentTitle = ({ title, titleDescription, buttonTitle, editForm }) 
   //   setModal(true)
   // }
 
+  const checkModal = (title) => {};
 
   return (
     <>
@@ -43,13 +51,19 @@ const PageComponentTitle = ({ title, titleDescription, buttonTitle, editForm }) 
 
         {title == "PARTICIPANT" ? (
           <>
-            {" "}
             <ParticipantModal modal={modal} setModal={setModal} />
+          </>
+        ) : title == "LEVEL" ? (
+          <>
+            <LevelModal modal={modal} setModal={setModal} />
+          </>
+        ) : title == "MODULE" ? (
+          <>
+            <ModuleModal modal={modal} setModal={setModal} />
           </>
         ) : (
           <>
-            {" "}
-            <Modal modal={modal} setModal={setModal} />
+            <OrganizationModal modal={modal} setModal={setModal} />
           </>
         )}
       </div>
