@@ -4,15 +4,13 @@ import PureModal from "react-pure-modal";
 import "react-pure-modal/dist/react-pure-modal.min.css";
 import { useForm } from "react-hook-form";
 import { SERVER_LINK } from "../../helpers/config";
-import axios from 'axios';
+import axios from "axios";
 import { useRouter } from "next/router";
 
-
-const ParticipantModal = ({ modal, setModal ,editForm , participantId}) => {
+const ParticipantModal = ({ modal, setModal, editForm, participantId }) => {
   //For Image Preview
   const [selectedImage, setSelectedImage] = useState();
   const router = useRouter();
-
 
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
@@ -22,35 +20,29 @@ const ParticipantModal = ({ modal, setModal ,editForm , participantId}) => {
 
   const [password, setPassword] = useState("");
   const [organizationId, setOrganizationId] = useState("");
-  
+
   const { register, handleSubmit } = useForm();
 
-//   console.log('this is the modeal calle');
+  //   console.log('this is the modeal calle');
 
-
-  if(editForm){
-
-      axios
+  if (editForm) {
+    axios
       .get(`${SERVER_LINK}/participants/${participantId}`)
       .then((response) => {
-   
         let singleParticipantData = response.data;
 
         setName(singleParticipantData.name);
         setEmail(singleParticipantData.email);
         setMobile(singleParticipantData.mobile);
-        setOrganizationId(singleParticipantData.Organization_id)
-        setPassword(singleParticipantData.password)
-        setButtonText('Update')
+        setOrganizationId(singleParticipantData.Organization_id);
+        setPassword(singleParticipantData.password);
+        setButtonText("Update");
       })
       .catch((err) => {
         console.log(err);
       });
-
   }
 
-  
- 
   // This function will be triggered when the file field change
   const imageChange = (e) => {
     if (e.target.files && e.target.files.length > 0) {
@@ -63,34 +55,29 @@ const ParticipantModal = ({ modal, setModal ,editForm , participantId}) => {
     setSelectedImage();
   };
 
-
-    
-const handleEditClick = (participantId) => {
-    // setOpen(true); 
+  const handleEditClick = (participantId) => {
+    // setOpen(true);
     setButtonText("Update");
     // setEditForm(true);
     setParticipantId(participantId);
-    console.log("participant id "+participantId)
+    console.log("participant id " + participantId);
 
     // first find the user with the id
     axios
       .get(`${SERVER_LINK}/participants/${participantId}`)
       .then((response) => {
-   
         let singleParticipantData = response.data;
 
         setName(singleParticipantData.name);
         setEmail(singleParticipantData.email);
         setMobile(singleParticipantData.mobile);
-        setOrganizationId(singleParticipantData.Organization_id)
-        setPassword(singleParticipantData.password)
-       
+        setOrganizationId(singleParticipantData.Organization_id);
+        setPassword(singleParticipantData.password);
       })
       .catch((err) => {
         console.log(err);
       });
   };
-
 
   useEffect(() => {
     if (!modal) {
@@ -100,11 +87,14 @@ const handleEditClick = (participantId) => {
 
   // for sending the data to the backend
   const checkWithDatabase = async (data) => {
-    console.log('This is thge data ');
-    console.log(data);
-    // data.status = true;
-    data = JSON.stringify(data);
-    console.log(data);
+    data.name = name
+  data.email = email 
+  data.mobile = mobile
+  data.id = organizationId
+  data.password = password
+
+  // data.status = true;
+  let participantData = JSON.stringify(data);
 
     // for taking the patch api data
     if (editForm) {
@@ -116,7 +106,7 @@ const handleEditClick = (participantId) => {
           },
         })
         .then((response) => {
-       setModal(!modal)
+          setModal(!modal);
           router.replace(router.asPath);
         })
         .catch((err) => {
@@ -133,11 +123,16 @@ const handleEditClick = (participantId) => {
           Accept: "application/json",
           "Content-Type": "application/json;charset=UTF-8",
         },
-        data,
+        data : participantData,
       })
         .then((response) => {
           router.replace(router.asPath);
-          setModal(!modal)
+          setName("")
+          setEmail("")
+          setMobile("")
+          setPassword("")
+          setOrganizationId("")
+          setModal(!modal);
         })
         .catch((err) => {
           console.log(err);
@@ -152,6 +147,11 @@ const handleEditClick = (participantId) => {
         isOpen={modal}
         width="800px"
         onClose={() => {
+          setName("")
+          setEmail("")
+          setMobile("")
+          setPassword("")
+          setOrganizationId("")
           setModal(false);
           return true;
         }}
@@ -179,9 +179,8 @@ const handleEditClick = (participantId) => {
                     id="name"
                     type="text"
                     value={name}
-                 {...register("name", {
-                onChange: (e) => setName(e.target.value),
-              })}
+                    onChange = { (e) => setName(e.target.value)}
+                    required = "required"
                     placeholder="Jane"
                   />
                   {/* <p className="text-red-500 text-xs italic">
@@ -200,10 +199,9 @@ const handleEditClick = (participantId) => {
                     id="email"
                     type="email"
                     placeholder="example@gmail.com "
+                    required = "required"
                     value={email}
-                    {...register("email", {
-                   onChange: (e) => setEmail(e.target.value),
-                 })}
+                    onChange = { (e) => setEmail(e.target.value)}
                   />
                 </div>
               </div>
@@ -221,10 +219,9 @@ const handleEditClick = (participantId) => {
                     id="password"
                     type="password"
                     placeholder="******************"
+                    required = "required"
                     value={password}
-                    {...register("password", {
-                   onChange: (e) => setPassword(e.target.value),
-                 })}
+                    onChange = { (e) => setPassword(e.target.value)}
                   />
                   <p className="text-gray-600 text-xs italic">
                     Make it as long and as crazy as you'd like
@@ -232,8 +229,6 @@ const handleEditClick = (participantId) => {
                 </div>
               </div>
 
-            
-              
               <div className="flex flex-wrap -mx-3 mb-6">
                 <div className="w-full md:w-1/2 px-3 mb-6 md:mb-0">
                   <label
@@ -247,10 +242,9 @@ const handleEditClick = (participantId) => {
                     id="mobile"
                     type="text"
                     placeholder="+91 "
+                    required = "required"
                     value={mobile}
-                    {...register("mobile", {
-                   onChange: (e) => setMobile(e.target.value),
-                 })}
+                    onChange = { (e) => setMobile(e.target.value)}
                   />
                   {/* <p className="text-red-500 text-xs italic">
                     Please fill out this field.   property - > border-red-500
@@ -268,22 +262,25 @@ const handleEditClick = (participantId) => {
                     id="org_id"
                     type="text"
                     placeholder="e.g. 1000"
+                    required = "required"
                     value={organizationId}
-                    {...register("Organization_id", {
-                        onChange: (e) => setOrganizationId(e.target.value),
-                      })}
+                    onChange = { (e) => setOrganizationId(e.target.value)}
                   />
                 </div>
-                
               </div>
-              <button type="submit" className="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">{buttonText}</button>
+              <button
+                type="submit"
+                className="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+              >
+                {buttonText}
+              </button>
             </form>
           </div>
 
           {/* */}
         </div>
       </PureModal>
-      ;
+      
     </>
   );
 };
