@@ -1,24 +1,23 @@
 // import Table from 'rc-table';
-import Table from './Table';
-import React, { useState } from 'react';
+import Table from "./Table";
+import React, { useState } from "react";
 import Pagination from "react-js-pagination";
-import axios from 'axios';
+import axios from "axios";
 import { SERVER_LINK } from "../../helpers/config";
 import { useRouter } from "next/router";
-import Modal from '../common/Modal';
+import Modal from "../common/Modal";
 import PureModal from "react-pure-modal";
 import "react-pure-modal/dist/react-pure-modal.min.css";
-import { useForm } from "react-hook-form";  
+import { useForm } from "react-hook-form";
 
-
-const OrganizationTable = ({org_data}) => {
+const OrganizationTable = ({ org_data }) => {
   // console.log('this is the talbe ');
-  
+
   const router = useRouter();
   const [editForm, setEditForm] = useState(false);
-  const [modal,setModal] = useState(false);
-  const [organizationId,setOrganizationId] = useState("")
-  const [orgData,setOrgData] = useState();
+  const [modal, setModal] = useState(false);
+  const [organizationId, setOrganizationId] = useState("");
+  const [orgData, setOrgData] = useState();
   const [buttonText, setButtonText] = useState("Add");
 
   const [name, setName] = useState("");
@@ -29,8 +28,9 @@ const OrganizationTable = ({org_data}) => {
   const [state, setState] = useState("");
   const [mobile, setMobile] = useState("");
   const [quota, setQuota] = useState("");
+
   // const [buttonText, setButtonText] = useState("Add");
-  
+
   const [password, setPassword] = useState("");
 
   const { register, handleSubmit } = useForm();
@@ -48,15 +48,43 @@ const OrganizationTable = ({org_data}) => {
       });
   };
 
-  const handleEditClick = (org_id) => {
+  const handleBoxClick = async (org_id ,org_status) =>{
+    console.log('This is hte box click');
+    console.log(org_id);
+    let new_status = {
+      status : ! org_status
+    }
+    new_status = JSON.stringify(new_status)
+    console.log(new_status);
+    
+    
+    await axios
+        .patch(`${SERVER_LINK}/rest-api/${org_id}`, new_status, {
+          headers: {
+            Accept: "application/json",
+            "Content-Type": "application/json;charset=UTF-8",
+          },
+        })
+        .then((response) => {
+          // setModal(!modal);
+          router.replace(router.asPath);
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+
+    
+    
+  }
+  const handleEditClick = async (org_id) => {
     // setOpen(true);
     setButtonText("Update");
     setEditForm(true);
     setOrganizationId(org_id);
-    setModal(true)
+    setModal(true);
 
     // first find the user with the id
-    axios
+    await axios
       .get(`${SERVER_LINK}/rest-api/${org_id}`)
       .then((response) => {
         let singleOrgData = response.data;
@@ -65,6 +93,7 @@ const OrganizationTable = ({org_data}) => {
         setEmail(singleOrgData.email);
         setMobile(singleOrgData.mobile);
         setState(singleOrgData.state);
+        setPassword(singleOrgData.password);
         setAddress(singleOrgData.address);
         setCity(singleOrgData.city);
         setPincode(singleOrgData.pincode);
@@ -76,20 +105,30 @@ const OrganizationTable = ({org_data}) => {
   };
 
   const checkWithDatabase = async (data) => {
-    data.status = true;
-    data = JSON.stringify(data);
+    // data.status = true;
+    data.name = name;
+    data.email = email;
+    data.mobile = mobile;
+    data.password = password;
+    data.city = city;
+    data.state = state;
+    data.pincode = pincode;
+    data.address = address;
+    data.quota = quota;
+    let OrganizationData = JSON.stringify(data);
+    console.log(data);
 
     // for taking the patch api data
     if (editForm) {
       await axios
-        .patch(`${SERVER_LINK}/rest-api/${organizationId}`, data, {
+        .patch(`${SERVER_LINK}/rest-api/${organizationId}`, OrganizationData, {
           headers: {
             Accept: "application/json",
             "Content-Type": "application/json;charset=UTF-8",
           },
         })
         .then((response) => {
-         setModal(!modal)
+          setModal(!modal);
           router.replace(router.asPath);
         })
         .catch((err) => {
@@ -109,7 +148,7 @@ const OrganizationTable = ({org_data}) => {
         data,
       })
         .then((response) => {
-          setModal(!modal)
+          setModal(!modal);
           router.replace(router.asPath);
         })
         .catch((err) => {
@@ -118,7 +157,7 @@ const OrganizationTable = ({org_data}) => {
     }
   };
 
-  function createData(name, email, org_id) {
+  function createData(name, email, org_id, org_status) {
     const action = (
       <>
         <button
@@ -138,12 +177,19 @@ const OrganizationTable = ({org_data}) => {
     );
     const status = (
       <>
-     <div className="flex justify-evenly">
-  {/* <div className="form-check form-switch"> */}
-    <input className="form-check-input appearance-none w-9 -ml-10 rounded-full float-left h-5 align-top bg-white bg-no-repeat bg-contain bg-gray-300 focus:outline-none cursor-pointer shadow-sm" type="checkbox" role="switch" id="flexSwitchCheckDefault"/>
-    {/* <label className="form-check-label inline-block text-gray-800" for="flexSwitchCheckDefault">Default switch checkbox input</label> */}
-  {/* </div> */}
-</div>
+        <div className="flex ">
+          {/* <div className="form-check form-switch"> */}
+          <input
+            onClick = {() => handleBoxClick(org_id,org_status)}
+            className="form-check-input appearance-none w-9  rounded-full float-left h-5 align-top bg-white bg-no-repeat bg-contain bg-gray-300 focus:outline-none cursor-pointer shadow-sm"
+            type="checkbox"
+            role="switch"
+            id="flexSwitchCheckDefault"
+            defaultChecked = {org_status}
+          />
+          {/* <label className="form-check-label inline-block text-gray-800" for="flexSwitchCheckDefault">Default switch checkbox input</label> */}
+          {/* </div> */}
+        </div>
       </>
     );
     return { name, email, status, action };
@@ -153,77 +199,82 @@ const OrganizationTable = ({org_data}) => {
     let name = element.name;
     let email = element.email;
     let org_id = element.id;
-    return createData(name, email, org_id);
+    let org_status = element.status;
+    return createData(name, email, org_id, org_status);
   });
-  
-    const columns = [
-        {
-          Header: "Name",
-          accessor: 'name',
-          title: 'Name',
-          dataIndex: 'name',
-          key: 'name',
-          width: 400,
-          className:"text-white bg-gray-800 p-2 border-r-2 border-b-2",
-          rowClassName:"bg-black-ripon"
-        },
-        {
-          Header: "Email",
-          accessor: 'email',
-          title: 'Email',
-          dataIndex: 'email',
-          key: 'email',
-          width: 400,
-          className:"text-white bg-gray-600 p-2 border-r-2 border-b-2"
-        },
-        {  
-          Header: "Status",
-          accessor: 'status',
-          title: 'Status',
-          dataIndex: 'status',
-          key: 'status',
-          width: 400,
-          className:"text-white bg-gray-800 p-2 border-r-2 border-b-2",
-        //   render : () =><>  <div class="flex justify-center">
-        //   <div class="form-check form-switch">
-        //     <input class="form-check-input appearance-none w-9 -ml-10 rounded-full float-left h-5 align-top bg-black bg-no-repeat   focus:outline-none cursor-pointer shadow-sm" type="checkbox" role="switch" id="flexSwitchCheckChecked" checked/>          </div>
-        // </div></>
-        },
-        {
-          Header: "Action",
-          accessor: 'action',
-          title: 'Action',
-          dataIndex: 'action',
-          key: 'operations',
-          width:250,
-          className:"text-white bg-gray-600 p-2 border-b-2",
-//           render: (id) => <><button class="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded-full">
-//           Edit
-//         </button> <button  class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-full"  >
-//   Delete
-// </button></>,
-          
-        },
-      ];
-      
-      // const data = [
-      //   { id:'01', name: 'Jack', email: 28 },
-      //   { id:'02', name: 'Rose', email: 36 },
-      // ];
 
-      // data by using which table data is creating using api call
-      const data = rowsDataArray;
+  const columns = [
+    {
+      Header: "Name",
+      accessor: "name",
+      title: "Name",
+      dataIndex: "name",
+      key: "name",
+      width: 400,
+      className: "text-white bg-gray-800 p-2 border-r-2 border-b-2",
+      rowClassName: "bg-black-ripon",
+    },
+    {
+      Header: "Email",
+      accessor: "email",
+      title: "Email",
+      dataIndex: "email",
+      key: "email",
+      width: 400,
+      className: "text-white bg-gray-600 p-2 border-r-2 border-b-2",
+    },
+    {
+      Header: "Status",
+      accessor: "status",
+      title: "Status",
+      dataIndex: "status",
+      key: "status",
+      width: 400,
+      className: "text-white bg-gray-800 p-2 border-r-2 border-b-2",
+      //   render : () =><>  <div class="flex justify-center">
+      //   <div class="form-check form-switch">
+      //     <input class="form-check-input appearance-none w-9 -ml-10 rounded-full float-left h-5 align-top bg-black bg-no-repeat   focus:outline-none cursor-pointer shadow-sm" type="checkbox" role="switch" id="flexSwitchCheckChecked" checked/>          </div>
+      // </div></>
+    },
+    {
+      Header: "Action",
+      accessor: "action",
+      title: "Action",
+      dataIndex: "action",
+      key: "operations",
+      width: 250,
+      className: "text-white bg-gray-600 p-2 border-b-2",
+      //           render: (id) => <><button class="bg-green-500 hover:bg-green-700 text-white font-bold py-2 px-4 rounded-full">
+      //           Edit
+      //         </button> <button  class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-full"  >
+      //   Delete
+      // </button></>,
+    },
+  ];
 
-      //Pagination
-      const [activePage, setActivePage] = useState(15)
-      const handlePageChange = (pageNumber)=>{
-        setActivePage(pageNumber)
-      }
+  // const data = [
+  //   { id:'01', name: 'Jack', email: 28 },
+  //   { id:'02', name: 'Rose', email: 36 },
+  // ];
 
-    return (
-        <>
-        <Table columns={columns} data={data} rowKey="id"  className='bg-white table-auto p-1 w-full text-center rc-table-custom font-semibold hover:table-fixed'/>
-        {/* <Pagination
+  // data by using which table data is creating using api call
+  const data = rowsDataArray;
+
+  //Pagination
+  const [activePage, setActivePage] = useState(15);
+  const handlePageChange = (pageNumber) => {
+    setActivePage(pageNumber);
+  };
+
+  return (
+    <>
+      <Table
+        columns={columns}
+        data={data}
+        rowKey="id"
+        className="bg-white table-auto p-1 w-full text-center rc-table-custom font-semibold hover:table-fixed"
+      />
+      {/* <Pagination
           activePage={activePage}
           itemsCountPerPage={10}
           totalItemsCount={450}
@@ -237,9 +288,9 @@ const OrganizationTable = ({org_data}) => {
           itemClass='js-li'
           linkClass='page-link'
         /> */}
-            {/* <Modal modal={modal} setModal={setModal} editForm ={editForm} orgData={orgData} organizationId={organizationId} /> */}
+      {/* <Modal modal={modal} setModal={setModal} editForm ={editForm} orgData={orgData} organizationId={organizationId} /> */}
 
-            <PureModal
+      <PureModal
         //header={<div className="bg-blue-600 p-2 font-bold text-lg text-center text-white">Category</div>}
         isOpen={modal}
         width="800px"
@@ -271,11 +322,13 @@ const OrganizationTable = ({org_data}) => {
                     id="grid-first-name"
                     type="text"
                     value={name}
+                    required
                     {...register("name", {
-                      onChange: (e) => setName(e.target.value)
+                      onChange: (e) => setName(e.target.value),
                     })}
                     placeholder="Jane"
                   />
+
                   {/* <p className="text-red-500 text-xs italic">
                     Please fill out this field.   property - > border-red-500
                   </p> */}
@@ -291,6 +344,7 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-email"
                     type="email"
+                    required
                     placeholder="example@gmail.com "
                     value={email}
                     {...register("email", {
@@ -312,12 +366,14 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-password"
                     type="password"
+                    required
                     placeholder="******************"
                     value={password}
                     {...register("password", {
                       onChange: (e) => setPassword(e.target.value),
                     })}
                   />
+
                   <p className="text-gray-600 text-xs italic">
                     Make it as long and as crazy as you'd like
                   </p>
@@ -336,6 +392,7 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-city"
                     type="text"
+                    required
                     placeholder="Albuquerque"
                     value={city}
                     {...register("city", {
@@ -355,6 +412,7 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-state"
                     type="text"
+                    required
                     placeholder="State"
                     value={state}
                     {...register("state", {
@@ -373,6 +431,7 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-zip"
                     type="text"
+                    required
                     placeholder="90210"
                     value={pincode}
                     {...register("pincode", {
@@ -393,6 +452,7 @@ const OrganizationTable = ({org_data}) => {
                     className="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
                     id="grid-address"
                     type="text"
+                    required
                     placeholder="Your Office number... "
                     value={address}
                     {...register("address", {
@@ -414,11 +474,13 @@ const OrganizationTable = ({org_data}) => {
                     id="grid-mobile"
                     type="text"
                     placeholder="+91 "
+                    required
                     value={mobile}
                     {...register("mobile", {
                       onChange: (e) => setMobile(e.target.value),
                     })}
                   />
+
                   {/* <p className="text-red-500 text-xs italic">
                     Please fill out this field.   property - > border-red-500
                   </p> */}
@@ -435,6 +497,7 @@ const OrganizationTable = ({org_data}) => {
                     id="grid-quota"
                     type="text"
                     placeholder="e.g. 1000"
+                    required
                     value={quota}
                     {...register("quota", {
                       onChange: (e) => setQuota(e.target.value),
@@ -454,10 +517,8 @@ const OrganizationTable = ({org_data}) => {
           {/* */}
         </div>
       </PureModal>
-
-        </>
-        
-    );
+    </>
+  );
 };
 
 export default OrganizationTable;
